requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.0
numpy==1.25.2
matplotlib==3.8.0
discord-webhook==1.3.0
openai==1.5.0
ccxt==4.0.0
python-dateutil==2.8.2
ta==0.10.2
schedule==1.2.0
pytest==7.4.0
